Plug unimplemented Automounter into main loop.
The Automounter object is supposed to handle device insertion and
removal as detected by our FdEvents object. It shall attempt to mount
the devices, unmount them, put them into a device database (an
internal AllDevices object), and do the D-Bus announcements.

This means it is also going to implement any mounting policies and
other details: What should be mounted, what not? Is it a USB device?
Where is it connected?

Change-Id: Id32109ff0478b0eed8d6bd4777777c205283cb8c