Generalize and fix Devices::Device::contains_volume().
The function was looking up the device by label, not by device name.
Bad naming of functions.

Instead of a boolean that tells if a volume with given device name is
already known, the volume pointer itself is returned. This makes the
function a lot more useful and avoid introduction of a lookup function
in addition to a test function.

Change-Id: I5da68c9d4e9e1ace82f24546bb0c8207c2b7c192