First half-working implementation of automounter.
Automatic mounting of USB devices should work (relies on external
tools and permission to use them). Directories are created and removed
as needed.

Volume information are not retrieved at the moment. Mounted volumes
are not gathered at startup (important for restarts after update or
crash). D-Bus method for retrieving the list of devices and mounted
volumes is not implemented yet.

Change-Id: Id7ffbd2ed1fa6475d24896e2de3c8230c9e0ab74