Unmount volumes and cleanup on startup (#141).
This is the simple way of doing a scan and reconstruction of mount
state on startup. After unmounting and cleaning up everything, any
mass storage device is going to be mounted in the next step.

This unmount/mount step has several bad side effects, but since this
case will almost never be hit in practice, this is going to be OK.

Change-Id: Ia3f95ea81b33a5db933a20883608a5fa07509c91